Remove encodec
